move LinkInfo and convert to dataclass
